+++ src/app/(beforeLogin)/_component/Footer.tsx
@@ -1,4 +1,4 @@
-import FooterLinkWrapper from "./FooterImgFrame";
+import FooterLinkWrapper from "./FooterLinkWrapper";
 import ImageWrapper from "./ImageWrapper";
 
 export default function Footer() {
